Fixes a server crash on claim event
Fixes a server crash when a specific type of claim is done, where X is
negative and Z is positive. If you make a 3x3 box, with the center and
upper left corner removed it would crash.

Also added protections so events like this will log but no longer crash
the server.

Resolves #3